tidy up analysis cached to reuse historical fires fetch, remove redundant alerts fetches
