+++ app/javascript/data/analysis-datasets.json
@@ -97,7 +97,37 @@
     "GLAD_GEOSTORE_SUMMARY": "2a22a8fc-5260-446e-b1f1-3695e325af7d",
     "GLAD_GEOSTORE_WEEKLY": "633e64ff-27d3-4203-942b-c7a93e632b45",
     "GLAD_GEOSTORE_DAILY": "69f279f5-c5e3-4856-bd8b-2b5eceaef56a",
-    "GLAD_GEOSTORE_WHITELIST": "b4f6dd8e-d8f4-4de3-ad67-441ea6d15977"
+    "GLAD_GEOSTORE_WHITELIST": "b4f6dd8e-d8f4-4de3-ad67-441ea6d15977",
+    "VIIRS_ADM0_WHITELIST": "86751f56-de4e-44f7-b07a-44b0fb861821",
+    "VIIRS_ADM1_WHITELIST": "b899ab6c-9982-4d32-b80e-04e68c47ebdb",
+    "VIIRS_ADM2_WHITELIST": "5f3d5866-4490-480e-ac27-f2861930c0dc",
+    "VIIRS_ADM0_WEEKLY": "17ff682d-6ef2-4934-9a0e-8dddb95e5c45",
+    "VIIRS_ADM1_WEEKLY": "0ebb725e-42e1-4a1f-8875-930d6f51113c",
+    "VIIRS_ADM2_WEEKLY": "8836395d-5173-45b8-bc3c-9dce271feb31",
+    "VIIRS_ADM2_DAILY": "6e451fd5-5241-451b-872f-27f7b87e6a5d",
+    "VIIRS_WDPA_SUMMARY": "",
+    "VIIRS_WDPA_WHITELIST": "197a7702-eafd-4ae7-8285-f14b86b676ff",
+    "VIIRS_WDPA_WEEKLY": "9b984feb-2d16-4414-9bf2-08235e3f3cc5",
+    "VIIRS_WDPA_DAILY": "c8040782-9dbb-4f6a-a8f8-51d43afdf81a",
+    "VIIRS_GEOSTORE_SUMMARY": "",
+    "VIIRS_GEOSTORE_WHITELIST": "4e0cd805-2339-41ce-af02-8ab093ce7ebc",
+    "VIIRS_GEOSTORE_WEEKLY": "290fa76b-d954-4473-bef2-b7e89f778436",
+    "VIIRS_GEOSTORE_DAILY": "5d7305db-e074-4e33-b011-c34dc09709ba",
+    "MODIS_ADM0_WHITELIST": "8c1cbe0c-61b0-4722-8c2f-2c0d81f8a8cc",
+    "MODIS_ADM1_WHITELIST": "a619de63-1ada-427c-9538-42a362f6a977",
+    "MODIS_ADM2_WHITELIST": "677bf3af-b500-4bea-91ae-4b84a9d9f9fa",
+    "MODIS_ADM0_WEEKLY": "58fee933-81bd-45bd-a1eb-6a37972e7c28",
+    "MODIS_ADM1_WEEKLY": "aae20051-c216-4b42-96d0-e8befc34f1cf",
+    "MODIS_ADM2_WEEKLY": "03e08f73-135a-464d-9ec3-ded08e6c69dd",
+    "MODIS_ADM2_DAILY": "7dfa052a-155a-4159-b3ba-d0f81152662e",
+    "MODIS_WDPA_SUMMARY": "",
+    "MODIS_WDPA_WHITELIST": "7b7ea04e-03f8-4ed6-9a2a-53e1c793af4a",
+    "MODIS_WDPA_WEEKLY": "a1e38d65-1e1f-4e2c-ba66-f7c449a82a9a",
+    "MODIS_WDPA_DAILY": "5be51b58-43f7-4973-aebf-46ae11ac62ba",
+    "MODIS_GEOSTORE_SUMMARY": "",
+    "MODIS_GEOSTORE_WHITELIST": "fab92867-45f0-4e47-af4e-73ea6b9372e3",
+    "MODIS_GEOSTORE_WEEKLY": "5586eb14-8016-4630-8e2b-7636a0b26983",
+    "MODIS_GEOSTORE_DAILY": "c0254765-e921-45e5-9eed-72b7294ce150"
   },
   "production": {
     "ANNUAL_ADM0_SUMMARY": "4ddbda72-b203-4049-8458-97acc2b00ec5",
@@ -132,6 +162,36 @@
     "GLAD_GEOSTORE_SUMMARY": "2a22a8fc-5260-446e-b1f1-3695e325af7d",
     "GLAD_GEOSTORE_WEEKLY": "633e64ff-27d3-4203-942b-c7a93e632b45",
     "GLAD_GEOSTORE_DAILY": "69f279f5-c5e3-4856-bd8b-2b5eceaef56a",
-    "GLAD_GEOSTORE_WHITELIST": "b4f6dd8e-d8f4-4de3-ad67-441ea6d15977"
+    "GLAD_GEOSTORE_WHITELIST": "b4f6dd8e-d8f4-4de3-ad67-441ea6d15977",
+    "VIIRS_ADM0_WHITELIST": "86751f56-de4e-44f7-b07a-44b0fb861821",
+    "VIIRS_ADM1_WHITELIST": "b899ab6c-9982-4d32-b80e-04e68c47ebdb",
+    "VIIRS_ADM2_WHITELIST": "5f3d5866-4490-480e-ac27-f2861930c0dc",
+    "VIIRS_ADM0_WEEKLY": "17ff682d-6ef2-4934-9a0e-8dddb95e5c45",
+    "VIIRS_ADM1_WEEKLY": "0ebb725e-42e1-4a1f-8875-930d6f51113c",
+    "VIIRS_ADM2_WEEKLY": "8836395d-5173-45b8-bc3c-9dce271feb31",
+    "VIIRS_ADM2_DAILY": "6e451fd5-5241-451b-872f-27f7b87e6a5d",
+    "VIIRS_WDPA_SUMMARY": "",
+    "VIIRS_WDPA_WHITELIST": "197a7702-eafd-4ae7-8285-f14b86b676ff",
+    "VIIRS_WDPA_WEEKLY": "9b984feb-2d16-4414-9bf2-08235e3f3cc5",
+    "VIIRS_WDPA_DAILY": "c8040782-9dbb-4f6a-a8f8-51d43afdf81a",
+    "VIIRS_GEOSTORE_SUMMARY": "",
+    "VIIRS_GEOSTORE_WHITELIST": "4e0cd805-2339-41ce-af02-8ab093ce7ebc",
+    "VIIRS_GEOSTORE_WEEKLY": "290fa76b-d954-4473-bef2-b7e89f778436",
+    "VIIRS_GEOSTORE_DAILY": "5d7305db-e074-4e33-b011-c34dc09709ba",
+    "MODIS_ADM0_WHITELIST": "8c1cbe0c-61b0-4722-8c2f-2c0d81f8a8cc",
+    "MODIS_ADM1_WHITELIST": "a619de63-1ada-427c-9538-42a362f6a977",
+    "MODIS_ADM2_WHITELIST": "677bf3af-b500-4bea-91ae-4b84a9d9f9fa",
+    "MODIS_ADM0_WEEKLY": "58fee933-81bd-45bd-a1eb-6a37972e7c28",
+    "MODIS_ADM1_WEEKLY": "aae20051-c216-4b42-96d0-e8befc34f1cf",
+    "MODIS_ADM2_WEEKLY": "03e08f73-135a-464d-9ec3-ded08e6c69dd",
+    "MODIS_ADM2_DAILY": "7dfa052a-155a-4159-b3ba-d0f81152662e",
+    "MODIS_WDPA_SUMMARY": "",
+    "MODIS_WDPA_WHITELIST": "7b7ea04e-03f8-4ed6-9a2a-53e1c793af4a",
+    "MODIS_WDPA_WEEKLY": "a1e38d65-1e1f-4e2c-ba66-f7c449a82a9a",
+    "MODIS_WDPA_DAILY": "5be51b58-43f7-4973-aebf-46ae11ac62ba",
+    "MODIS_GEOSTORE_SUMMARY": "",
+    "MODIS_GEOSTORE_WHITELIST": "fab92867-45f0-4e47-af4e-73ea6b9372e3",
+    "MODIS_GEOSTORE_WEEKLY": "5586eb14-8016-4630-8e2b-7636a0b26983",
+    "MODIS_GEOSTORE_DAILY": "c0254765-e921-45e5-9eed-72b7294ce150"
   }
 }
